navbar connecting auth routes
