+++ src/components/config/Apollo.js~HEAD
@@ -1,4 +1,3 @@
-import { ApolloProvider } from '@apollo/react-common';
 import { ApolloClient } from 'apollo-client';
 import { createHttpLink } from 'apollo-link-http';
 import { InMemoryCache } from "apollo-cache-inmemory";
