+++ src/ssr-server.js
@@ -17,7 +17,7 @@ import { config } from "dotenv";
 config({ path: path.resolve(__dirname, "../.env") });
 
 import mongoose from "mongoose";
-import { createTokens } from "./backend/middleware/auth";
+import { createTokens, authenticate } from "./backend/middleware/auth";
 
 mongoose.connect(process.env.MONGODB_URI, {
   useUnifiedTopology: true,
@@ -48,47 +48,7 @@ app
     const server = express();
 
     server.use(cookieParser());
-
-    server.use(async (req, res, next) => {
-      const refreshToken = req.cookies["refresh-token"];
-      const accessToken = req.cookies["access-token"];
-      if (!refreshToken && !accessToken) {
-        return next();
-      }
-
-      try {
-        const data = jwt.verify(accessToken, process.env.JWT_SECRET_KEY);
-        req.userId = data.userId;
-        return next();
-      } catch {}
-
-      if (!refreshToken) {
-        return next();
-      }
-
-      let data;
-
-      try {
-        data = jwt.verify(refreshToken, process.env.JWT_SECRET_KEY);
-      } catch {
-        return next();
-      }
-
-      const user = await User.findById(data.userId);
-
-      // token has been invalidated
-      if (!user || user.count !== data.count) {
-        return next();
-      }
-
-      const tokens = createTokens(user);
-
-      res.cookie("refresh-token", tokens.refreshToken);
-      res.cookie("access-token", tokens.accessToken);
-      req.userId = user.id;
-
-      next();
-    });
+    server.use(authenticate);
 
     apolloServer.applyMiddleware({ app: server, path: "/api" });
